Rename Analysis download button file
