+++ frontend/src/components/MapView/Legends/AnalysisDownloadButton.tsx
